Bring old `mdx` parser back
