+++ package.json
@@ -99,8 +99,11 @@
     "postcss-selector-parser": "2.2.3",
     "postcss-values-parser": "2.0.1",
     "regexp-util": "1.2.2",
+    "remark-footnotes": "2.0.0",
     "remark-math": "5.1.1",
+    "remark-math-v3": "3.0.1",
     "remark-parse": "10.0.1",
+    "remark-parse-v8": "npm:remark-parse@8.0.3",
     "sdbm": "2.0.0",
     "string-width": "5.0.1",
     "strip-ansi": "7.0.1",
